+++ src/utils/exit.ts
@@ -1,12 +1,13 @@
 /**
  * Exit utils.
  *
- * When calling process.exit() in worker thread used for file generation,
+ * When calling process.exit() in worker thread used for test-file generation,
  * logs are not flushed (https://github.com/vitalets/playwright-bdd/issues/59).
  * That's why instead of process.exit we throw ExitError
  * that just sets process.exitCode = 1 and allow program to exit normally.
+ * This esnured by wrapping code with withExitHandler().
  *
- * On the other hand, when running in main thread, especially inside Playwright,
+ * On the other hand, when running in the main thread, especially inside Playwright,
  * thrown error is captured by Playwright and show with additional messages (e.g. no tests found).
  * That's why in main thread we to call process.exit() to show only needed error.
  *
@@ -16,7 +17,6 @@
  * - https://github.com/cucumber/cucumber-js/pull/123
  */
 
-import { logger } from './logger';
 import { isMainThread } from 'node:worker_threads';
 
 class ExitError extends Error {
@@ -30,6 +30,8 @@ export async function withExitHandler(fn: () => unknown) {
     return await fn();
   } catch (e) {
     if (e instanceof ExitError) {
+      // eslint-disable-next-line no-console, max-depth
+      if (e.message) console.error(e.message);
       process.exitCode = 1;
     } else {
       throw e;
@@ -40,7 +42,10 @@ export async function withExitHandler(fn: () => unknown) {
 export function exit(...messages: string[]) {
   messages = messages.filter(Boolean);
   if (isMainThread) {
-    if (messages.length) logger.error('Error:', ...messages);
+    // use console.error() here instead of logger.error() to have less stack
+    // for flushing messages to stderr.
+    // eslint-disable-next-line no-console, max-depth
+    if (messages.length) console.error('Error:', ...messages);
     process.exit(1);
   } else {
     throw new ExitError(messages.join(' '));